requirements.txt
certifi==2021.5.30
chardet==4.0.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
numpy==1.20.3
Pillow==8.2.0
proglog==0.1.9
python-dotenv==0.17.1
pytube==10.8.4
requests==2.25.1
tqdm==4.61.1
urllib3==1.26.5
